[!!!][TASK] Enable local storage per default
Up to now per default bookmarks could only be added
when a frontend user was logged in. Now this has been
removed.

**Corrective action:**
In case the bookmarking functionality should remain to be
available for logged in users the TS constant
`plugin.tx_bookmarkpages.settings.storeLocal` should be
set to `0`. It is as well available in the constant editor.